# [optimize](string) optimize instr and locate function for constant arguments