# Proposed changes
Optimize instr and locate function for constant arguments.
1. instr and locate function constant arguments has 58%~200% performance improvement.
2. refactor locate(substr, str, pos) as standardized arguments processing.

(1) instr function has up 58%~200%

| needle size |  before | after | up |
|--------|--------|--------| --------|
| size 1 | 2.17s ![image](https://user-images.githubusercontent.com/67053339/232171238-3a18b81d-0208-4ac3-b379-f339f3c00e15.png) | 1.37s  ![image](https://user-images.githubusercontent.com/67053339/232171315-2752f147-b1f4-4900-b23e-979bdac1339b.png) |  58% |
| size 6 | 1.32s ![image](https://user-images.githubusercontent.com/67053339/232171257-f9776f15-ddd8-45f5-b09c-154c8a352560.png) | 0.66s  ![image](https://user-images.githubusercontent.com/67053339/232171286-7d968c78-9c24-4c7f-b6e3-e744ce8f6d46.png) |  200%  |

(2) locate function implement  is the same as instr, just arguments positions are different.



Issue Number: close #xxx

## Problem summary

Describe your changes.

## Checklist(Required)

* [ ] Does it affect the original behavior
* [ ] Has unit tests been added
* [ ] Has document been added or modified
* [ ] Does it need to update dependencies
* [ ] Is this PR support rollback (If NO, please explain WHY)

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...